updated front end tests
